requirements.txt
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
chromadb==0.4.24
sentence-transformers==2.7.0
requests==2.31.0
streamlit==1.33.0
python-dotenv==1.0.1
tiktoken==0.6.0  # (Optional, for token counting)
langchain-text-splitters